Remove panic/recover approach and replace with explicit error checking.
Recover was unexpectedly recovering from network-level errors that
should be regarded as fatal. Rather than whitelist specific errors, it
seems more idiomatic to always check for errors and return them to callers.

http://stackoverflow.com/a/18772200/245018 provides a detailed overview
of the advantages and disadvantages of "repetitive" error handling in
Go. In this commit I've adopted the convention which minimises the lines
of code while still maintaing explicit error handling in each function.